wip: app.js cleanup/refactor
