+++ elixir/web-ng/assets/js/hooks/index.js
@@ -1,0 +1,33 @@
+import BulkEditTagsToggle from "./BulkEditTagsToggle"
+import JdmEditorHook from "./JdmEditorHook"
+import LocalTime from "./LocalTime"
+import MapboxFlowMap from "./MapboxFlowMap"
+import SRQLTimeCookie from "./SRQLTimeCookie"
+import SpatialSurveyMap from "./SpatialSurveyMap"
+import NetflowSankeyChart from "./charts/NetflowSankeyChart"
+import NetflowStackedAreaChart from "./charts/NetflowStackedAreaChart"
+import NetflowStacked100Chart from "./charts/NetflowStacked100Chart"
+import NetflowLineSeriesChart from "./charts/NetflowLineSeriesChart"
+import NetflowGridChart from "./charts/NetflowGridChart"
+import GodViewBinaryStream from "./god_view/GodViewBinaryStream"
+import GodViewControlsState from "./god_view/GodViewControlsState"
+import TimeseriesChart from "./charts/TimeseriesChart"
+import TimeseriesCombinedChart from "./charts/TimeseriesCombinedChart"
+
+export default {
+  BulkEditTagsToggle,
+  JdmEditorHook,
+  LocalTime,
+  MapboxFlowMap,
+  SRQLTimeCookie,
+  SpatialSurveyMap,
+  NetflowSankeyChart,
+  NetflowStackedAreaChart,
+  NetflowStacked100Chart,
+  NetflowLineSeriesChart,
+  NetflowGridChart,
+  GodViewBinaryStream,
+  GodViewControlsState,
+  TimeseriesChart,
+  TimeseriesCombinedChart,
+}
